Fixing review findings
