+++ grpcutil/cancel.go
@@ -9,7 +9,6 @@ import (
 	"errors"
 
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 // IsCanceled checks whether an error comes from an operation being canceled
@@ -17,7 +16,7 @@ func IsCanceled(err error) bool {
 	if errors.Is(err, context.Canceled) {
 		return true
 	}
-	s, ok := status.FromError(err)
+	s, ok := ErrorToStatus(err)
 	if ok && s.Code() == codes.Canceled {
 		return true
 	}
