+++ grpcutil/metadata_test.go
@@ -7,24 +7,25 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/metadata"
-
-	"github.com/grafana/dskit/httpgrpc"
 )
 
 func TestAppendMessageSizeToOutgoingContext(t *testing.T) {
 	ctx := context.Background()
 
-	req := &httpgrpc.HTTPRequest{
-		Method: "GET",
-		Url:    "/test",
-	}
+	sizer := fakeSizer(5)
 
-	ctx = AppendMessageSizeToOutgoingContext(ctx, req)
+	ctx = AppendMessageSizeToOutgoingContext(ctx, sizer)
 
 	md, exists := metadata.FromOutgoingContext(ctx)
 	require.True(t, exists)
 
 	vals := md.Get(MetadataMessageSize)
 	require.Len(t, vals, 1)
-	require.Equal(t, strconv.Itoa(req.Size()), vals[0])
+	require.Equal(t, strconv.Itoa(sizer.Size()), vals[0])
+}
+
+type fakeSizer int
+
+func (f fakeSizer) Size() int {
+	return int(f)
 }
