+++ httpgrpc/httpgrpc_test.go
@@ -105,62 +105,6 @@ func TestHTTPResponseFromError(t *testing.T) {
 	}
 }
 
-func TestStatusFromError(t *testing.T) {
-	msgErr := "this is an error"
-	testCases := map[string]struct {
-		err            error
-		expectedStatus *status.Status
-	}{
-		"no error cannot be cast to status.Status": {
-			err: nil,
-		},
-		"a random error cannot be cast to status.Status": {
-			err: fmt.Errorf(msgErr),
-		},
-		"a wrapped error of a random error cannot be cast to status.Status": {
-			err: fmt.Errorf("wrapped: %w", fmt.Errorf(msgErr)),
-		},
-		"a gRPC error built by gogo/status can be cast to status.Status": {
-			err:            status.Error(codes.Internal, msgErr),
-			expectedStatus: status.New(codes.Internal, msgErr),
-		},
-		"a wrapped error of a gRPC error built by gogo/status can be cast to status.Status": {
-			err:            fmt.Errorf("wrapped: %w", status.Error(codes.Internal, msgErr)),
-			expectedStatus: status.New(codes.Internal, msgErr),
-		},
-		"a gRPC error built by grpc/status can be cast to status.Status": {
-			err:            grpcstatus.Error(codes.Internal, msgErr),
-			expectedStatus: status.New(codes.Internal, msgErr),
-		},
-		"a wrapped error of a gRPC error built by grpc/status can be cast to status.Status": {
-			err:            fmt.Errorf("wrapped: %w", grpcstatus.Error(codes.Internal, msgErr)),
-			expectedStatus: status.New(codes.Internal, msgErr),
-		},
-		"a gRPC error built by httpgrpc can be cast to status.Status": {
-			err:            Errorf(400, msgErr),
-			expectedStatus: status.New(400, msgErr),
-		},
-		"a wrapped gRPC error built by httpgrpc can be cast to status.Status": {
-			err:            fmt.Errorf("wrapped: %w", Errorf(400, msgErr)),
-			expectedStatus: status.New(400, msgErr),
-		},
-	}
-	for testName, testData := range testCases {
-		t.Run(testName, func(t *testing.T) {
-			stat, ok := statusFromError(testData.err)
-			if testData.expectedStatus == nil {
-				require.False(t, ok)
-				require.Nil(t, stat)
-			} else {
-				require.True(t, ok)
-				require.NotNil(t, stat)
-				require.Equal(t, testData.expectedStatus.Code(), stat.Code())
-				require.Equal(t, testData.expectedStatus.Message(), stat.Message())
-			}
-		})
-	}
-}
-
 func checkDetailAsHTTPResponse(t *testing.T, httpResponse *HTTPResponse, stat *status.Status) {
 	details := stat.Details()
 	require.Len(t, details, 1)
